+++ cmd/root.go
@@ -20,6 +20,7 @@ import (
 	"strings"
 
 	"github.com/astarte-platform/astartectl/cmd/appengine"
+	"github.com/astarte-platform/astartectl/cmd/cluster"
 	"github.com/astarte-platform/astartectl/cmd/housekeeping"
 	"github.com/astarte-platform/astartectl/cmd/pairing"
 	"github.com/astarte-platform/astartectl/cmd/realm"
@@ -70,6 +71,7 @@ func init() {
 	rootCmd.AddCommand(realm.RealmManagementCmd)
 	rootCmd.AddCommand(utils.UtilsCmd)
 	rootCmd.AddCommand(appengine.AppEngineCmd)
+	rootCmd.AddCommand(cluster.ClusterCmd)
 }
 
 // initConfig reads in config file and ENV variables if set.